requirements.txt
celery>=4.3.0,<4.4.0
flask>=1.1.4,<1.2.0
flask-cors>=3.0.10,<3.1.0
gunicorn>=20.1.0,<20.2.0
python-dotenv>=0.19.0,<0.20.0
pytz==2021.1
redis>=3.5.3,<3.6.0